refactor - interfaces as a separate file
